Refactor worker and driver
