+++ controller/controller.go
@@ -40,7 +40,7 @@ type unit struct {
 
 type baseController struct {
 	conf       *config.Config
-	newDriver  func(*config.Config) driver.Driver
+	newDriver  func(*config.Config, driver.Task) (driver.Driver, error)
 	fileReader func(string) ([]byte, error)
 	units      []unit
 	watcher    watcher
@@ -74,13 +74,13 @@ func (ctrl *baseController) init(ctx context.Context) error {
 	units := make([]unit, 0, len(tasks))
 
 	for _, task := range tasks {
-		d := ctrl.newDriver(ctrl.conf)
-		if err := d.Init(ctx); err != nil {
-			log.Printf("[ERR] (ctrl) error initializing driver: %s", err)
+		log.Printf("[DEBUG] (ctrl) initializing task %q", task.Name)
+		d, err := ctrl.newDriver(ctrl.conf, task)
+		if err != nil {
 			return err
 		}
-		log.Printf("[DEBUG] (ctrl) initializing task %q", task.Name)
-		err := d.InitTask(task, true)
+
+		err = d.InitTask(ctx, true)
 		if err != nil {
 			log.Printf("[ERR] (ctrl) error initializing task %q: %s", task.Name, err)
 			return err
@@ -105,21 +105,30 @@ func (ctrl *baseController) init(ctx context.Context) error {
 	return nil
 }
 
-func newDriverFunc(conf *config.Config) (func(*config.Config) driver.Driver, error) {
+func InstallDriver(ctx context.Context, conf *config.Config) error {
 	if conf.Driver.Terraform != nil {
-		log.Printf("[INFO] (ctrl) setting up Terraform driver")
+		tfConf := *conf.Driver.Terraform
+		return driver.InstallTerraform(ctx, *tfConf.Path, *tfConf.WorkingDir)
+	}
+	return errors.New("unsupported driver")
+}
+
+func newDriverFunc(conf *config.Config) (
+	func(conf *config.Config, task driver.Task) (driver.Driver, error), error) {
+	if conf.Driver.Terraform != nil {
 		return newTerraformDriver, nil
 	}
 	return nil, errors.New("unsupported driver")
 }
 
-func newTerraformDriver(conf *config.Config) driver.Driver {
+func newTerraformDriver(conf *config.Config, task driver.Task) (driver.Driver, error) {
 	tfConf := *conf.Driver.Terraform
 	return driver.NewTerraform(&driver.TerraformConfig{
+		Task:              task,
 		Log:               *tfConf.Log,
 		PersistLog:        *tfConf.PersistLog,
 		Path:              *tfConf.Path,
-		WorkingDir:        *tfConf.WorkingDir,
+		WorkingDir:        filepath.Join(*tfConf.WorkingDir, task.Name),
 		Backend:           tfConf.Backend,
 		RequiredProviders: tfConf.RequiredProviders,
 		ClientType:        *conf.ClientType,
